refactor: isola a função que lida com o checkbox
