+++ app/frontend/src/components/helpers/checkbox.js
@@ -1,0 +1,6 @@
+export const allChecked = () => {
+  const nodeListOfCheckboxInput = document.querySelectorAll('.checkbox')
+  const arrayOfCheckboxInput = [...nodeListOfCheckboxInput]
+
+  return arrayOfCheckboxInput
+}
