+++ app/frontend/src/components/helpers/display.js
